refactor(transport): move relay URL helpers into nostr-server utils
